[refactor] balance에서 point로 명칭 변경
